+++ meja/src/com/dua3/meja/model/Color.java
@@ -262,7 +262,7 @@ public class Color {
         if (s.startsWith("#")) {
             // FIXME JDK 8
             // int i = Integer.parseUnsignedInt(s.substring(1), 16);
-            int i = Integer.parseInt(s.substring(1), SHIFT_R);
+            int i = (int) Long.parseLong(s.substring(1), SHIFT_R);
             return new Color(i);
         }
